requirements.txt
fastapi
pymongo
python-dotenv
streamlit
streamlit-option-menu
pandas
matplotlib
python-jose[cryptography]
passlib[bcrypt]
python-multipart
uvicorn
Pillow
numpy
